refactor: move shared helper to helpers-file
