+++ core/src/commands/cloud/secrets/secrets-list.ts
@@ -6,9 +6,7 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-import queryString from "query-string"
 import { ConfigurationError } from "../../../exceptions.js"
-import type { ListSecretsResponse } from "@garden-io/platform-api-types"
 import { printHeader } from "../../../logger/util.js"
 import { dedent, deline, renderTable } from "../../../util/string.js"
 import type { CommandParams, CommandResult } from "../../base.js"
@@ -16,31 +14,10 @@ import { Command } from "../../base.js"
 import { applyFilter, noApiMsg } from "../helpers.js"
 import { sortBy } from "lodash-es"
 import { StringsParameter } from "../../../cli/params.js"
-import type { CloudApi, CloudProject } from "../../../cloud/api.js"
-import type { Log } from "../../../logger/log-entry.js"
+import type { CloudProject } from "../../../cloud/api.js"
 import { styles } from "../../../logger/styles.js"
 import type { SecretResult } from "./secret-helpers.js"
-import { makeSecretFromResponse } from "./secret-helpers.js"
-
-export const fetchAllSecrets = async (api: CloudApi, projectId: string, log: Log): Promise<SecretResult[]> => {
-  let page = 0
-  const secrets: SecretResult[] = []
-  let hasMore = true
-  while (hasMore) {
-    log.debug(`Fetching page ${page}`)
-    const q = queryString.stringify({ projectId, offset: page * pageLimit, limit: pageLimit })
-    const res = await api.get<ListSecretsResponse>(`/secrets?${q}`)
-    if (res.data.length === 0) {
-      hasMore = false
-    } else {
-      secrets.push(...res.data.map((secret) => makeSecretFromResponse(secret)))
-      page++
-    }
-  }
-  return secrets
-}
-
-const pageLimit = 100
+import { fetchAllSecrets } from "./secret-helpers.js"
 
 export const secretsListOpts = {
   "filter-envs": new StringsParameter({
